+++ docs/conf.py
@@ -406,6 +406,7 @@ intersphinx_mapping = {
     'sphinx_nbexamples_doc': (
         'http://sphinx-nbexamples.readthedocs.io/en/latest/', None),
     'psyplot': ('http://psyplot.readthedocs.io/en/latest/', None),
+    'nbconvert': ('https://nbconvert.readthedocs.io/en/latest/', None),
 }
 if six.PY3:
     intersphinx_mapping['python'] = ('https://docs.python.org/3.4/', None)